# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for ecommerce website",
        "title": "Zog_festiv eCommerce API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/addapparel": {
            "post": {
                "description": "Adding new product of category apparel in database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Product\u0026Offer Management"
                ],
                "summary": "Adding new product",
                "parameters": [
                    {
                        "description": "Apparel Data",
                        "name": "admin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ApparelInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Apparel"
                        }
                    }
                }
            }
        },
        "/addcoupon": {
            "post": {
                "description": "Addig coupon for users, with a unique code",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin Product\u0026Offer Management"
                ],
                "summary": "Adding coupon by admin",
                "parameters": [
                    {
                        "description": "coupon",
                        "name": "admin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Coupon"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success masage",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/addoffer": {
            "post": {
                "description": "Addig coupon for users, with a unique code",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin Product\u0026Offer Management"
                ],
                "summary": "Adding offer by admin",
                "parameters": [
                    {
                        "description": "offer",
                        "name": "admin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Offer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success masage",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/addticket": {
            "post": {
                "description": "Adding new product of category ticket in database",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Product\u0026Offer Management"
                ],
                "summary": "Adding new product",
                "parameters": [
                    {
                        "description": "Ticket Data",
                        "name": "admin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.TicketInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Ticket"
                        }
                    }
                }
            }
        },
...
        "/adminlogin": {
            "post": {
                "description": "Admin login with otp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Authentication"
                ],
                "summary": "Admin login with otp",
                "parameters": [
                    {
                        "description": "Admin Data",
                        "name": "admin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Login"
                        }
                    }
                }
            }
        },
        "/adminloginpassword": {
            "post": {
                "description": "Admin login with password and phone number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Authentication"
                ],
                "summary": "Admin Login with password",
                "parameters": [
                    {
                        "description": "Admin Data",
                        "name": "admin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Login"
                        }
                    }
                }
            }
        },
        "/adminotpvalidation": {
            "post": {
                "description": "Otp Validation for admin login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Authentication"
                ],
                "summary": "Otp validation",
                "parameters": [
                    {
                        "description": "Admin Data",
                        "name": "admin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Login"
                        }
                    }
                }
            }
        },
...
        "/deleteapparel/{id}": {
            "delete": {
                "description": "Soft deleting the data of a product from database in category apparel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Product\u0026Offer Management"
                ],
                "summary": "Delete existing product from database",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "product id",
                        "name": "ProductId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Apparel"
                        }
                    }
                }
            }
        },
        "/deleteticket/{id}": {
            "delete": {
                "description": "Soft deleting the data of a product from database in category ticket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Product\u0026Offer Management"
                ],
                "summary": "Delete existing product from database",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "product id",
                        "name": "ProductId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Ticket"
                        }
                    }
                }
            }
        },
        "/editapparel/{id}": {
            "put": {
                "description": "Edit data of a product in category apparel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Product\u0026Offer Management"
                ],
                "summary": "Edit existing product data",
                "parameters": [
                    {
                        "description": "Apparel Data",
                        "name": "admin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Apparel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Apparel"
                        }
                    }
                }
            }
        },
...
        "/editticket/{id}": {
            "put": {
                "description": "Edit data of a product in category ticket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Product\u0026Offer Management"
                ],
                "summary": "Edit existing product data",
                "parameters": [
                    {
                        "description": "Ticket Data",
                        "name": "admin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Ticket"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Ticket"
                        }
                    }
                }
            }
        },
...
        "/registeradmin": {
            "post": {
                "description": "Adding new admin to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Authentication"
                ],
                "summary": "registering new admin",
                "parameters": [
                    {
                        "description": "Admin Data",
                        "name": "admin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Admin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Admin"
                        }
                    }
                }
            }
        },
...
        "/searchuser": {
            "get": {
                "description": "Showing user list for management by searching with name or id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin User Management"
                ],
                "summary": "Search user by id or name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "User Id",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.User"
                        }
                    }
                }
            }
        },
...
        "/sortuser": {
            "get": {
                "description": "Showing user list for management by sorting with permisiion",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin User Management"
                ],
                "summary": "User list by permission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "page no",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "limit no",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "true/false",
                        "name": "permission",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.User"
                        }
                    }
                }
            }
        },
...
        "/userpermission/{id}": {
            "post": {
                "description": "Toggling user permission for block/unblock",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin User Management"
                ],
                "summary": "block/unblock user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Admin"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.ApparelInput": {
            "type": "object",
            "properties": {
                "apparelid": {
                    "type": "integer"
                },
                "category": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "imageurl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "productCategory": {
                    "type": "string"
                },
                "productId": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "size": {
                    "type": "string"
                },
                "subcategory": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}